extractors/bilibili: Separate into a package
